Remove some errors

+++ Big-BWT/tfm_index_construct.cpp
@@ -128,7 +128,7 @@ int_vector<> load_parse(const string &infile) {
 
     int_vector<> v(n, 0, 32);
     uint32_t tmp, max = 0, min = (1 << 31) - 1;
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         fread(&tmp, sizeof(uint32_t), 1, parse);
         //tmp--; // the lowest phrase ID is 1 // not anymore
         v[i] = tmp;
